pyproject.toml
[project]
name = "home-assistant-ziroom"
version = "0.1.0"
description = "Home Assistant integration for Ziroom smart devices"
requires-python = ">=3.9"

[project.optional-dependencies]
dev = [
    "python-dotenv>=1.0.0",
]
